Consolidate IHttpHandler endpoint creation to ensure delegate is set
The RequestDelegate was not being set for one of the pathways that creates endpoints from IHttpHandler. This change moves all the creation of Endpoint/EndpointBuilder into IHttpHandlerEndpointFactory so that it is in one place and will be set the same for all endpoints.
