+++ src/Handlers/SystemWeb/HttpHandlerEndpointConventionBuilder.cs
@@ -5,26 +5,17 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.AspNetCore.Routing.Patterns;
-using Microsoft.AspNetCore.SystemWebAdapters;
 using Microsoft.AspNetCore.SystemWebAdapters.HttpHandlers;
 using Microsoft.Extensions.Primitives;
 
 namespace System.Web;
 
-internal sealed class HttpHandlerEndpointConventionBuilder : EndpointDataSource, IEndpointConventionBuilder
+internal sealed class HttpHandlerEndpointConventionBuilder(
+    IEnumerable<IHttpHandlerCollection> managers,
+    IHttpHandlerEndpointFactory factory) : EndpointDataSource, IEndpointConventionBuilder
 {
-    private readonly IHttpHandlerCollection[] _managers;
-    private readonly RequestDelegate _defaultHandler;
-
+    private readonly IHttpHandlerCollection[] _managers = [.. managers];
     private List<Action<EndpointBuilder>> _conventions = [];
-
-    public HttpHandlerEndpointConventionBuilder(
-        IEnumerable<IHttpHandlerCollection> managers,
-        IServiceProvider services)
-    {
-        _managers = managers.ToArray();
-        _defaultHandler = BuildDefaultHandler(services);
-    }
 
     public override IReadOnlyList<Endpoint> Endpoints
     {
@@ -35,7 +26,7 @@ internal sealed class HttpHandlerEndpointConventionBuilder : EndpointDataSource,
             foreach (var (route, metadataCollection) in CollectMetadata())
             {
                 var pattern = RoutePatternFactory.Parse(route);
-                var builder = new RouteEndpointBuilder(_defaultHandler, pattern, 0);
+                var builder = factory.CreateBuilder(pattern);
 
                 builder.AddHandler(metadataCollection);
 
@@ -44,12 +35,10 @@ internal sealed class HttpHandlerEndpointConventionBuilder : EndpointDataSource,
                     convention(builder);
                 }
 
-#if NET7_0_OR_GREATER
                 if (builder.FilterFactories.Count > 0)
                 {
                     throw new NotSupportedException("Filter factories are not supported for handlers");
                 }
-#endif
 
                 endpoints.Add(builder.Build());
             }
@@ -89,25 +78,6 @@ internal sealed class HttpHandlerEndpointConventionBuilder : EndpointDataSource,
         => (_conventions ??= []).Add(convention);
 
     public override IChangeToken GetChangeToken() => new CompositeChangeToken(_managers.Select(m => m.GetChangeToken()).ToArray());
-
-    private static RequestDelegate BuildDefaultHandler(IServiceProvider services)
-    {
-        var builder = new ApplicationBuilder(services);
-
-        builder.EnsureRequestEndThrows();
-        builder.Run(context =>
-        {
-            if (context.AsSystemWeb().CurrentHandler is { } handler)
-            {
-                return handler.RunHandlerAsync(context).AsTask();
-            }
-
-            context.Response.StatusCode = 500;
-            return context.Response.WriteAsync("Invalid handler");
-        });
-
-        return builder.Build();
-    }
 
     private sealed class MappedHandlerMetadata(string route, IHttpHandlerMetadata metadata) : IHttpHandlerMetadata
     {
